feat: make judger.py easier to use
Add a verbose arg to cli, default to false and will NOT display incorrect output (diff content) in stdout, add -v or --verbose to enable verbosed output to see detailed difference. Change sorting for testcases to int(filename), for example, 2.in comes after 1.in not 10.in previously. Assign colors for different results utilizing ANSI, by default, a RE(runtime error) is red, a WA(wrong answer) is yellow, and an AC(answer correct) is green, you can change them in global constants in the front of judger file.